**IMPORTANT** I am committing the MainMenuHUD with the default widget set to the scenario menu, this is only to make testing the menus easier. This default widget needs to be changed back to the MainMenuWidget
This committ includes the DataRecorder menu with a large amount of functionality, it can display a list of data recording points matching a name, data type and boolean toggle. It cannot yet return to the parent menu.

I also have found a way to sync up all of the background images in all the menu which use them, as such I have updated these menus with this fix and committed them. These are the ScenarioMenuWidget and all of the List Menus.